Refactored worldgen (again). x/y/zBlock in IPosBlock are now Ints (I caved.) Nyx worldgen will be refined to take advantage of the new system later.

+++ src/main/scala/mod/iceandshadow3/lib/compat/world/impl/AModDimension.java
@@ -144,9 +144,9 @@ public class AModDimension extends ModDimension {
 		public BlockPos findSpawn(int x, int z, boolean checkValid) {
 			return doFindSpawn(new IPosColumn() {
 				@Override
-				public long xBlock() { return x; }
+				public int xBlock() { return x; }
 				@Override
-				public long zBlock() { return z; }
+				public int zBlock() { return z; }
 			}, checkValid);
 		}
 
@@ -181,12 +181,12 @@ public class AModDimension extends ModDimension {
 		public boolean doesXZShowFog(int x, int z) {
 			return dimlogic.hasFogAt(new IPosColumn() {
 				@Override
-				public long xBlock() {
+				public int xBlock() {
 					return x;
 				}
 
 				@Override
-				public long zBlock() {
+				public int zBlock() {
 					return z;
 				}
 			});
