+++ src/main/scala/mod/iceandshadow3/lib/spatial/IPosTypes.scala
@@ -3,7 +3,7 @@ package mod.iceandshadow3.lib.spatial
 import net.minecraft.util.math.BlockPos
 
 object IPosChunk {
-	val CHUNK_BITS: Long = 4
+	val CHUNK_BITS: Int = 4
 	val CHUNK_MULT: Long = 1L << CHUNK_BITS
 	val CHUNK_MASK: Long = CHUNK_MULT - 1
 }
@@ -14,10 +14,10 @@ trait IPosChunk {
 
 trait IPosColumn extends IPosChunk {
 	import IPosChunk._
-	def xBlock: Long
-	def zBlock: Long
-	def xChunk: Int = (xBlock >> CHUNK_BITS).toInt
-	def zChunk: Int = (zBlock >> CHUNK_BITS).toInt
+	def xBlock: Int
+	def zBlock: Int
+	def xChunk: Int = xBlock >> CHUNK_BITS
+	def zChunk: Int = zBlock >> CHUNK_BITS
 	def xSubChunk: Byte = (xBlock & CHUNK_MASK).toByte
 	def zSubChunk: Byte = (zBlock & CHUNK_MASK).toByte
 }
